+++ src/App.js
@@ -2,6 +2,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import JobSearchPage from './pages/JobSearchPage';
 import SearchPage from './pages/SearchPage';
 import InterviewExperience from './pages/InterviewExperience';
+import JobDetail from './pages/JobDetail';
 import 'antd/dist/antd.min.css';
 import './App.css';
 import React from "react";
@@ -14,6 +15,7 @@ function App() {
 				<Route path='/login' element = {<SearchPage/>}/>
 				<Route path='/InterviewExperience' element = {<InterviewExperience/>}/>
 				<Route path='/jobSearch' element={<JobSearchPage />} />
+				<Route path='/jobDetail/:id' element={<JobDetail />} />
 			</Routes>
 		</BrowserRouter>
 	);
