+++ src/utils/searchEngine.js
@@ -1,0 +1,5 @@
+import {addHistory} from './indexDb';
+export function doSearch(searchInput,setSearchResult) {
+	addHistory(searchInput);
+	setSearchResult([]);
+}
